+++ LootGoblin/Windows/ConfigWindow.cs
@@ -11,11 +11,10 @@ public class ConfigWindow : Window, IDisposable
 
     public ConfigWindow(Plugin plugin) : base("Loot Goblin Settings###LootGoblinConfig")
     {
-        Flags = ImGuiWindowFlags.NoResize | ImGuiWindowFlags.NoCollapse | ImGuiWindowFlags.NoScrollbar |
-                ImGuiWindowFlags.NoScrollWithMouse;
+        Flags = ImGuiWindowFlags.NoCollapse;
 
-        Size = new Vector2(300, 150);
-        SizeCondition = ImGuiCond.Always;
+        Size = new Vector2(350, 220);
+        SizeCondition = ImGuiCond.FirstUseEver;
 
         configuration = plugin.Configuration;
     }
@@ -38,6 +37,26 @@ public class ConfigWindow : Window, IDisposable
     {
         ImGui.Text("Loot Goblin Settings");
         ImGui.Separator();
+        ImGui.Spacing();
+
+        var enabled = configuration.Enabled;
+        if (ImGui.Checkbox("Bot Enabled", ref enabled))
+        {
+            configuration.Enabled = enabled;
+            configuration.Save();
+        }
+
+        var showMain = configuration.ShowMainWindow;
+        if (ImGui.Checkbox("Show Main Window on Login", ref showMain))
+        {
+            configuration.ShowMainWindow = showMain;
+            configuration.Save();
+        }
+
+        ImGui.Spacing();
+        ImGui.Separator();
+        ImGui.Spacing();
+        ImGui.Text("UI Settings");
         ImGui.Spacing();
 
         var movable = configuration.IsConfigWindowMovable;